Add convenience and "raw" versions of all methods
Each public method now has a version prepended with an underscore; those
versions return an axios response directly as a promise, and the client
can choose to muck around with it.

The non-underscore versions make use of a helper function `extract()`
which simply extracts and returns the `.data` element of the axios
response in a new promise, and fails with the error response when the
axios response fails. That allows for try-catch handling of these
functions in client code.